==> notebooks/1.Exploration_of_Direction_API.ipynb <==
# Exploring Direction API


this method is used to find information about Direction API response data. This data exploration is used to get an overview of the data that can be used to make a better and more accurate model.

import json
import requests
import numpy as np
from sklearn.metrics.pairwise import cosine_similarity
from scipy.spatial.distance import cosine
from geopy import distance

API_KEY = '[API_KEY]'
LAT_START, LONG_START = -7.871300, 112.527282
LAT_END, LONG_END =-7.896923, 112.534245

response = requests.get(f'https://maps.googleapis.com/maps/api/directions/json?origin={LAT_START},{LONG_START}&destination={LAT_END},{LONG_END}&key={API_KEY}')

data = response.json()
routes = data['routes'][0]['legs'][0]['steps']
routes

cb_dumps = json.dumps(routes)
print(len(cb_dumps))

cb_load = json.loads(cb_dumps)
cb_load

data

LAT_ROUTES = []
LONG_ROUTES = []

for route in routes:
  lat_start = route['start_location']['lat']
  long_start = route['start_location']['lng']
  lat_end = route['end_location']['lat']
  long_end = route['end_location']['lng']

  if lat_start not in LAT_ROUTES:
    LAT_ROUTES.append(lat_start)
  if long_start not in LONG_ROUTES:
    LONG_ROUTES.append(long_start)
  if lat_end not in LAT_ROUTES:
    LAT_ROUTES.append(lat_end)
  if long_end not in LONG_ROUTES:
    LONG_ROUTES.append(long_end)

print(LAT_ROUTES)
print(LONG_ROUTES)

print(np.array(LAT_ROUTES).shape)
print(np.array(LONG_ROUTES).shape)

LAT_ROUTES = np.array(LAT_ROUTES)
LONG_ROUTES = np.array(LONG_ROUTES)

model = LinearRegression()
model = model.fit(LAT_ROUTES.reshape(-1,1), LONG_ROUTES)
print(model.score(LAT_ROUTES.reshape(-1,1), LONG_ROUTES))
print(model.coef_) # Gradient
print(model.intercept_) # Bias

# berati dia persamaannya adalah LONG_ROUTES = LAT_ROUTES*(-0.233..) + 110.69..
# ATAU JADI LAT_ROUTES*(0.233..) + LONG_ROUTES - 110.69.. = 0
# Gradient adalah -0.233..

# Summary

Data that must be sent from driver

*   Lat + Long Start (Required) 
*   Lat + Long End (Required)
*   Capacity (Required)
*   Time (Required)
*   Way point 1 Lat + Long (Opt)
*   Way point 2 Lat + Long (Opt)


def getRoutes(START_POINT, END_POINT, API_KEY): # WAY POINT 1 SAMA WAY POINT 2
  LAT_START, LONG_START = START_POINT
  LAT_END, LONG_END = END_POINT

  response = requests.get(f'https://maps.googleapis.com/maps/api/directions/json?origin={LAT_START},{LONG_START}&destination={LAT_END},{LONG_END}&key={API_KEY}')
  data = response.json()
  routes = data['routes'][0]['legs'][0]['steps']

  LAT_ROUTES = []
  LONG_ROUTES = []
  
  for route in routes:
    lat_start = route['start_location']['lat']
    long_start = route['start_location']['lng']
    lat_end = route['end_location']['lat']
    long_end = route['end_location']['lng']

    if lat_start not in LAT_ROUTES:
      LAT_ROUTES.append(lat_start)
    if long_start not in LONG_ROUTES:
      LONG_ROUTES.append(long_start)
    if lat_end not in LAT_ROUTES:
      LAT_ROUTES.append(lat_end)
    if long_end not in LONG_ROUTES:
      LONG_ROUTES.append(long_end)

  ROUTES_PAIR = [[LAT_ROUTES[i], LONG_ROUTES[i]] for i in range(len(LAT_ROUTES))]

  return ROUTES_PAIR

API_KEY = 'AIzaSyC9rKUqSrytIsC7QrPExD8v7oLNB3eOr5k'
START_POINT = -7.871300, 112.527282
END_POINT =-7.896923, 112.534245

getRoutes(START_POINT, END_POINT, API_KEY)

# Driver POint 0
DRIVER_START = -7.871300, 112.527282 # ALUN - ALUN
DRIVER_END =-7.896923, 112.534245 # BNS

# second driver point 1
SCND_START = -7.8823031,112.5298174 # PASAR BATU
SCND_END = -7.8346175,112.5267023 # Punten

# thrid driver point 2
THRD_START = -7.848781,112.5112838 # sumberejo
THRD_END =  -7.8783015,112.533646 # LIPPO

# First Passanger point 0
FP_START = -7.896923, 112.534245 # BNS
FP_END = -7.871300, 112.527282 # ALUN - ALUN

# Second Passanger Point 1
SP_START = -7.8792102,112.5279684 # SMA 1 BATU
SP_END = -7.8874847,112.5311676 # ORO - ORO OMBO

# Third Passanger Point 2
TP_START = -7.871300, 112.527282 # ALUN - ALUN
TP_END = -7.8186924,112.5270671 # SELECTA

# Fourth passanger point 3
FTP_START = -7.8694093,112.5285059 # SMP 2 Batu
FTP_END = -7.8935175,112.5540405 # JTP 3

# Fifth passanger point 4
FTH_START = -7.8786568,112.5321967 # LIPPO PLAZA
FTH_END = -7.8854482,112.5285224 # ORO - ORO OMBO

# six passanger point 5
XP_START = -7.8734831,112.5112557 # Suropati
XP_END = -7.8770156,112.5328306 # PARADIS

DRIVER = np.array(DRIVER_END) - np.array(DRIVER_START)
SCND = np.array(SCND_END) - np.array(SCND_START)
THRD = np.array(THRD_END) - np.array(THRD_START)
FP = np.array(FP_END) - np.array(FP_START)
SP = np.array(SP_END) - np.array(SP_START)
TP = np.array(TP_END) - np.array(TP_START)
FTP = np.array(FTP_END) - np.array(FTP_START)
FTH = np.array(FTH_END) - np.array(FTH_START)
XP = np.array(XP_END) - np.array(XP_START)

drv = [DRIVER, SCND, THRD]
psg = [FP, SP, TP, FTP, FTH, XP]

for i in range(len(drv)):
  for j in range(len(psg)):
    print(f'cosine distance from driver {i} with passanger {j} is : {cosine(drv[i], psg[j])}')
  print("=======================================================================================")

for i in range(len(drv)):
  for j in range(len(psg)):
    print(f'cosine similarity from driver {i} with passanger {j} is : {cosine_similarity([drv[i]], [psg[j]])[0][0]}')
  print("=======================================================================================")

print(f'cosine similarity from driver with passenger 1 is : {cosine_similarity([DRIVER], [FP])[0][0]}')
print(f'cosine similarity from driver with passenger 2 is : {cosine_similarity([DRIVER], [SP])[0][0]}')
print(f'cosine similarity from driver with passenger 3 is : {cosine_similarity([DRIVER], [TP])[0][0]}')
print(f'cosine similarity from driver with passenger 4 is : {cosine_similarity([DRIVER], [FTP])[0][0]}')
print(f'cosine similarity from driver with passenger 4 is : {cosine_similarity([DRIVER], [FTH])[0][0]}')

print(f'cosine similarity from driver 2 with passenger 1 is : {cosine_similarity([SCND], [FP])[0][0]}')
print(f'cosine similarity from driver 2 with passenger 2 is : {cosine_similarity([SCND], [SP])[0][0]}')
print(f'cosine similarity from driver 2 with passenger 3 is : {cosine_similarity([SCND], [TP])[0][0]}')
print(f'cosine similarity from driver 2 with passenger 4 is : {cosine_similarity([SCND], [FTP])[0][0]}')
print(f'cosine similarity from driver 2 with passenger 4 is : {cosine_similarity([SCND], [FTH])[0][0]}')

print(f'cosine distance from driver with passanger 1 is : {cosine(DRIVER, FP)}')
print(f'cosine distance from driver with passanger 2 is : {cosine(DRIVER, SP)}')
print(f'cosine distance from driver with passanger 3 is : {cosine(DRIVER, TP)}')
print(f'cosine distance from driver with passanger 4 is : {cosine(DRIVER, FTP)}')
print(f'cosine distance from driver with passanger 4 is : {cosine(DRIVER, FTH)}')

print(f'cosine distance from driver 2 with passanger 1 is : {cosine(SCND, FP)}')
print(f'cosine distance from driver 2 with passanger 2 is : {cosine(SCND, SP)}')
print(f'cosine distance from driver 2 with passanger 3 is : {cosine(SCND, TP)}')
print(f'cosine distance from driver 2 with passanger 4 is : {cosine(SCND, FTP)}')
print(f'cosine distance from driver 2 with passanger 4 is : {cosine(SCND, FTH)}')

basedriver = np.array(DRIVER_END) - np.array(DRIVER_START)
driverpickup = np.array(FTP_START) - np.array(DRIVER_START)
cosine_similarity([basedriver], [driverpickup])

driverpickup2 = np.array(SP_START) - np.array(DRIVER_START)
cosine_similarity([basedriver], [driverpickup2])

driverpickup3 = np.array(FTH_START) - np.array(DRIVER_START)
cosine_similarity([basedriver], [driverpickup3])

dist = distance.distance(FTP_START, DRIVER_START).km
dist

# coba waypoint
start = -7.8592855,112.5122669
via = [[-7.8692791,112.5115805],[-7.8794483,112.5277865]]
end = -7.8917706,112.5489331
response2 = requests.get(f'https://maps.googleapis.com/maps/api/directions/json?origin={start}&destination={end}&waypoints=via:{via[0][0]}%2C{via[0][1]}%7Cvia:{via[1][0]}%2C{via[1][1]}&departure_time=now&key={API_KEY}')
data2 = response2.json()
data2
