==> notebooks/2.Kotlite_Matching_Algorithm.ipynb <==
# KOTLITE Matching Algorithm

**Kotlite** (Angkot Elite) is an application that allows drivers to get passengers who have the same lane. This application is expected to parse existing congestion using the concept of ridesharing, in which passengers will get the experience of driving using a private car or taxi, but get a fairly cheap price similar to the price of public transportation. By using the machine learning algorithm, it is possible to match drivers and passengers who have the same routes.

in this case the dataset used is NYC Taxi trip duration obtained from [Kaggle](https://www.kaggle.com/debanjanpaul/new-york-city-taxi-trip-distance-matrix). In this dataset, there are pickup locations and dropoff locations that will try to be used to match drivers and passengers. Existing data will be manipulated and will be separated as driver data and passenger data.



import pandas as pd
import numpy as np
from geopy import distance
import datetime

import warnings
warnings.filterwarnings('ignore')

## Read the dataset

df = pd.read_csv('/content/drive/Shareddrives/Brillante Workspace/ML Corner/Dataset/NYC_dataset/train_distance_matrix.csv')
df.describe(include='all')

df.head().T

## Selected the features

def selected_features(dataframe):
  select_col = ['id', 'pickup_datetime', 'pickup_latitude', 'pickup_longitude',
                'dropoff_latitude', 'dropoff_longitude']
  return dataframe[select_col]

df = selected_features(df)
df.describe(include='all')

df.head()

df['pickup_datetime'] = pd.to_datetime(df['pickup_datetime'])
df.head()

dt = df.loc[0, 'pickup_datetime']
diff = datetime.timedelta(hours=1)
hasil = dt + diff
new_filter = df.loc[(df['pickup_datetime'] < hasil) & (df['pickup_datetime'] > dt)]
new_filter

## Split the dataset

The dataset will be separated by 20k for the dummy driver data and 5k for the dummy passenger data.

rd_dummy = df.sample(20000)
rd_dummy = rd_dummy.reset_index(drop=True)
ps_dummy = df.sample(5000)
ps_dummy = ps_dummy.reset_index(drop=True)

rd_dummy.head(10)

ps_dummy.head(10)

## Create Algorithm

The matching algorithm in Kotlite refers to the Nearest Neighbor algorithm. The reason we choose this algorithm is because the data is very dynamic when creating the application, so we need an algorithm model that can still match and recommend passengers who are looking for drivers who have the same route.

def l2_distances(point1, point2):
  l2 = 0
  for i in range(len(point1)):
    l2 += (point1[i] - point2[i]) ** 2
  return l2 ** 0.5

def l1_distances(point1, point2):
  l1 = 0
  for i in range(len(point1)):
    l1 += abs(point1[i] - point2[i])
  return l1

def nearest(data, query, k, dist='l2'):
  # query = [datetime, latpick, longpick, latdrop, longdrop]
  # get
  dt = query[0]
  diff = datetime.timedelta(minutes=30)
  up = dt + diff
  down = dt - diff
  data = data.loc[(data['pickup_datetime'] < up) & (data['pickup_datetime'] > down)]
  dt = data.to_numpy()

  distances_and_id = []

  if dist == 'l1':
    for values in dt:
      distance = l1_distances(values[2:], query[1:])
      distances_and_id.append((distance, values[0]))
  elif dist == 'l2':
    for values in dt:
      distance = l2_distances(values[2:], query[1:])
      distances_and_id.append((distance, values[0]))
  
  sorted_distances = sorted(distances_and_id)

  k_pickup_nearest = sorted_distances[:k]

  return k_pickup_nearest

test = ps_dummy.sample(1).to_numpy().reshape((-1))[1:]
test

recomendation = nearest(rd_dummy, test, 50, dist='l1')
recomendation

df[df['id'] == recomendation[1][1]].to_numpy().reshape((-1))

## Add another approach

def approaching2(start_point_driver, end_point_driver, 
                 pick_point_passenger, drop_point_passenger):
  # jarak tempuh
  jt_driver = distance.distance(start_point_driver, end_point_driver).km
  jt_passenger = distance.distance(pick_point_passenger, drop_point_passenger).km

  # radius from start and end point driver and passenger
  rad_start = distance.distance(start_point_driver, pick_point_passenger).km
  rad_end = distance.distance(end_point_driver, drop_point_passenger).km

  # logic
  # if jt_driver > jt_passenger:
  #   # recommend = True
  #   if (rad_start < 1) & (rad_end < 1):
  #     recommend = True
  #   else:
  #     recommend = False
  # else:
  #   if (rad_start < 1) & (rad_end < 1):
  #     recommend = True
  #   else:
  #     recommend = False

  if (rad_start < 2) & (rad_end < 2):
    recommend = True
  else:
    recommend = False
  return rad_start, rad_end, recommend

def recomendations(data, recommend, query):
  recommendation_by_range = []
  for rc in recommend:
    dt = data[data['id'] == rc[1]].to_numpy().reshape((-1))
    
    pass_pick = (query[1], query[2])
    pass_drop = (query[3], query[4])
    driv_start = (dt[2], dt[3])
    driv_end = (dt[4], dt[5])

    range_start, range_end, recommendation = approaching2(driv_start, driv_end, pass_pick, pass_drop)

    if recommendation == True:
      recommendation_by_range.append((range_start, range_end, dt[0]))
    
  sorted_by_range = sorted(recommendation_by_range)
  return sorted_by_range

recomendations(rd_dummy, recomendation, test)

df[df['id'] == 'id0213482'].to_numpy()

test

# Result

The test process is carried out using google maps. This is done to determine the level of effectiveness and accuracy of the algorithm in providing driver recommendations for the passanger. the purpose of this algorithm is to provide recommendations that are not burdensome from the driver's side.
