==> notebooks/3.Kotlite_Dataset_Maker.ipynb <==
# Kotelite dataset maker

**Kotlite** (Angkot Elite) is an application that allows drivers to get passengers who have the same lane. This application is expected to parse existing congestion using the concept of ridesharing, in which passengers will get the experience of driving using a private car or taxi, but get a fairly cheap price similar to the price of public transportation. By using the machine learning algorithm, it is possible to match drivers and passengers who have the same routes.

in this case the dataset used is NYC Taxi trip duration obtained from [Kaggle](https://www.kaggle.com/debanjanpaul/new-york-city-taxi-trip-distance-matrix). In this dataset, there are pickup locations and dropoff locations that will try to be used to match drivers and passengers. Existing data will be manipulated and will be separated as driver data and passenger data.

In this experiment, we want to use the response data from the direction API. the data we will use is the route points that the driver will pass. so that we can connect passengers who are around these points. by using this approach, we hope to get better recommendations and less burdensome to drivers.

import requests
import json
import pandas as pd
import numpy as np
import random
from datetime import datetime

import warnings
warnings.filterwarnings('ignore') 

## Read the dataset

df = pd.read_csv('/content/drive/Shareddrives/Brillante Workspace/ML Corner/Dataset/NYC_dataset/train_distance_matrix.csv')
df.describe(include='all')

## Selected the features

def filtering_dataset(dataframe):
  selected_ft = ['id', 'pickup_datetime', 'pickup_latitude', 'pickup_longitude',
                 'dropoff_latitude', 'dropoff_longitude']
  dataframe = dataframe[selected_ft]
  return dataframe

dfs = filtering_dataset(df)
dfs

# Split dataset

def change_second_dt(dataframe):
  for i in range(len(dataframe)):
    dtime = datetime.strptime(dataframe.loc[i,'pickup_datetime'], "%m/%D/%Y %H:%M")
    _delta = dtime - datetime(1970, 1, 1)
    dataframe.loc[i,'datetime'] = _delta.total_seconds()
  return dataframe 

def driver_and_passanger(dataframe, driver_sum=None, passanger_sum=None):
  df = dataframe.copy()
  if driver_sum == None:
    total = random.randint(10,500)
    driver_data = df.sample(total)
  else:
    driver_data = df.sample(driver_sum)
  
  if passanger_sum == None:
    total = random.randint(10,500)
    driver_data = df.sample(total)
  else:
    passanger_data = df.sample(passanger_sum)
  
  driver_data = driver_data.reset_index(drop=True)
  passanger_data = passanger_data.reset_index(drop=True)
  
  return driver_data, passanger_data

driver_dump, passanger_dump = driver_and_passanger(dfs, driver_sum=100, passanger_sum=1000)
driver_dump.describe(include='all')

passanger_dump.describe(include='all')

# Parsing Function

This function is made to get route data that will be passed by the driver. The data is obtained from the results of the response from the Google direction API.

def route_parsing(response):
  data = response.json()
  routes = data['routes'][0]['legs'][0]['steps']

  lat_routes = []
  long_routes = []

  for route in routes:
    lat_start = route['start_location']['lat']
    long_start = route['start_location']['lng']
    lat_end = route['end_location']['lat']
    long_end = route['end_location']['lng']

    if lat_start not in lat_routes:
      lat_routes.append(lat_start)
    if long_start not in long_routes:
      long_routes.append(long_start)
    if lat_end not in lat_routes:
      lat_routes.append(lat_end)
    if long_end not in long_routes:
      long_routes.append(long_end)
  
  routes_pair = [[lat_routes[i], long_routes[i]] for i in range(len(lat_routes))]
  return routes_pair

def get_routes(dataframe, API_KEY):
  df = dataframe.copy()

  for i in range(len(df)):
    start_lat = df.loc[i,'pickup_latitude']
    start_long = df.loc[i,'pickup_longitude']
    end_lat = df.loc[i,'dropoff_latitude']
    end_long = df.loc[i,'dropoff_longitude']

    # request data from Direction API
    response = requests.get(f'https://maps.googleapis.com/maps/api/directions/json?origin={start_lat},{start_long}&destination={end_lat},{end_long}&key={API_KEY}')
    # parse response to get routes data
    routes = route_parsing(response)
    # change list to string
    routes = json.dumps(routes)
    # pop data to dataframe
    df.loc[i, 'routes'] = routes
  return df

API_KEY = '[API KEY]'
driver_dump = get_routes(driver_dump, API_KEY)
driver_dump

# Save model to csv

driver_dump.to_csv('/content/drive/Shareddrives/Brillante Workspace/ML Corner/Dataset/kotlite_dataset/kotlite_driver_dataset.csv', index=False, header=True)
passanger_dump.to_csv('/content/drive/Shareddrives/Brillante Workspace/ML Corner/Dataset/kotlite_dataset/kotlite_passanger_dataset.csv', index=False, header=True)
