==> notebooks/4.Kotlite_Recommendation_Algorithm_(NYC_CASE).ipynb <==
# Recommendation Algorithm (NYC CASE)

at this point we want to create a recommendation system using the route data from the [Direction API](https://developers.google.com/maps/documentation/directions/overview). We will perform location-based filtering and direction-based filtering using the route data. This approach is aimed at further narrowing down the drivers that the system will recommend to prospective passengers. This system will also consider the direction of the driver and passenger whether in one direction or not.

import json
import numpy as np
import pandas as pd
from scipy.spatial.distance import cosine
from geopy import distance
import datetime

import warnings
warnings.filterwarnings('ignore')

# Read Dataset

driver_df = pd.read_csv('/content/drive/Shareddrives/Brillante Workspace/ML Corner/Dataset/kotlite_dataset/kotlite_driver_dataset.csv')
driver_df['pickup_datetime'] = pd.to_datetime(driver_df['pickup_datetime'])
driver_df.describe(include='all')

passanger_df = pd.read_csv('/content/drive/Shareddrives/Brillante Workspace/ML Corner/Dataset/kotlite_dataset/kotlite_passanger_dataset.csv')
passanger_df['pickup_datetime'] = pd.to_datetime(passanger_df['pickup_datetime'])
passanger_df.describe(include='all')

# create l1 distance function
def l1_distance(point1, point2):
  l1 = 0
  for i in range(len(point1)):
    l1 += abs(point1[i] - point2[i])
  return l1

# to get minimum point from start or end passanger point
def driver_data_handler(single_driv_data, passanger_point):
  # passanger point == [latitude, longitude]
  routes = single_driv_data[-1]

  driver_routes = []

  # pop start point data to driver routes
  driver_routes.append([single_driv_data[2], single_driv_data[3]])
  
  # parse routes and iterate to pop driver point to driver routes
  route_parse = json.loads(routes)
  for _route in route_parse:
    driver_routes.append(_route)
  
  # pop end point data to driver routes
  driver_routes.append([single_driv_data[4], single_driv_data[5]])
  
  # after get driver_routes list, calculate distance
  dist = []
  for driver_point in driver_routes:
    l1_dist = l1_distance(passanger_point, driver_point)
    dist.append(l1_dist)
  
  # get minimum values info
  min_val = min(dist)
  idx = dist.index(min_val)
  nearest_point = driver_routes[idx]
  return min_val, idx, nearest_point

# get nearest driver from passanger point
# query include [pickup_datetime, pickup_latitude, pickup_longitude, dropoff_latitude, dropoff_longitude]
def nearest1st_step(driver_data, query):
  # nearest 1st step is use for get nearest distance, and in one way
  picktime = query[0]
  
  # create filter for 30 minutes before and after picktime
  diff = datetime.timedelta(minutes=30)
  upth = picktime + diff
  lowth = picktime - diff
  
  # filter driver_data
  driver_data = driver_data.loc[(driver_data['pickup_datetime'] < upth) & (driver_data['pickup_datetime'] > lowth)]

  # get distance from routes point data
  dist = []
  for drdt in driver_data.values:
    start, _snear, pu_nearest_point = driver_data_handler(drdt, [query[1], query[2]])
    end, _enear, do_nearest_point = driver_data_handler(drdt, [query[3], query[4]])
    if _snear <= _enear:
      dist.append((start + end, drdt[0], pu_nearest_point, do_nearest_point))
  
  # sorting distance
  sorted_dist = sorted(dist)

  # list for recomendations
  recommendations = []
  # get real distance from pickup/dropoff point from snear or enear point
  for data in sorted_dist:
    # distance from pickup point to pu_nearest_point
    pick_distance = distance.distance([query[1], query[2]], data[2]).km
    drop_distance = distance.distance([query[3], query[4]], data[3]).km
    if (pick_distance <= 1.0) & (drop_distance <= 1.0).km:
      recommendations.append((data[1]))

  return sorted_dist

pass_data = passanger_df.sample(1).values.reshape(-1)
pass_data = pass_data[1:]
pass_data

nearest1st_step(driver_df, pass_data)

driver_df[driver_df['id'] == 'id3216829'].values
